Commit progress on BSDR API
Likely to dump API interface as these are not documented and more importantly, because there are limits to the number of records that can be drawn in any POST() call and this makes it much more difficult to put together a full response. Was also having trouble flattening the JSON response. Moving to using CSV files